Moved SmartGridWorldSimulation to its own file
